bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
distro==1.9.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
loguru==0.7.3
multidict==6.4.4
openai==1.82.0
propcache==0.3.1
pydantic==2.11.5
pydantic-core==2.33.2
pyrefly==0.17.0
python-dotenv==1.1.0
python-telegram-bot==22.1
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.1
yarl==1.20.0
